chore: rename link to ExtendedLink
